# Bulk load perf experiment