--- a/geom/alg_intersects.go
+++ b/geom/alg_intersects.go
@@ -205,6 +205,7 @@ func hasIntersectionBetweenLines(
 		}
 	}
 	tree := rtree.BulkLoad(bulk)
+	defer tree.Recycle()
 
 	// Keep track of an envelope of all of the points that are in the
 	// intersection.