--- a/geom/type_multi_polygon.go
+++ b/geom/type_multi_polygon.go
@@ -58,17 +58,8 @@ func validateMultiPolygon(polys []Polygon, opts ...ConstructorOption) error {
 	polyBoundaries := make([]indexedLines, len(polys))
 	polyBoundaryPopulated := make([]bool, len(polys))
 
-	bulkItems := make([]rtree.BulkItem, 0, len(polys))
-	for i, p := range polys {
-		env, ok := p.Envelope()
-		if ok {
-			item := rtree.BulkItem{Box: env.box(), RecordID: i}
-			bulkItems = append(bulkItems, item)
-		}
-	}
-	tree := rtree.BulkLoad(bulkItems)
+	var tree rtree.RTree
 	defer tree.Recycle()
-
 	for i := range polys {
 		env, ok := polys[i].Envelope()
 		if !ok {
@@ -77,10 +68,6 @@ func validateMultiPolygon(polys []Polygon, opts ...ConstructorOption) error {
 		box := env.box()
 
 		err := tree.RangeSearch(box, func(j int) error {
-			if i <= j {
-				return nil
-			}
-
 			for _, k := range [...]int{i, j} {
 				if !polyBoundaryPopulated[k] {
 					polyBoundaries[k] = newIndexedLines(polys[k].Boundary().asLines())
@@ -129,6 +116,8 @@ func validateMultiPolygon(polys []Polygon, opts ...ConstructorOption) error {
 		if err != nil {
 			return err
 		}
+
+		tree.Insert(box, i)
 	}
 	return nil
 }