--- a/rtree/bulk.go
+++ b/rtree/bulk.go
@@ -39,7 +39,9 @@ func calculateLevels(numItems int) int {
 
 func bulkInsert(items []BulkItem, levels int) *node {
 	if levels == 1 {
-		root := &node{isLeaf: true, numEntries: len(items)}
+		root := nodePool.Get().(*node)
+		root.isLeaf = true
+		root.numEntries = len(items)
 		for i, item := range items {
 			root.entries[i] = entry{
 				box:      item.Box,
@@ -76,11 +78,8 @@ func bulkInsert(items []BulkItem, levels int) *node {
 }
 
 func bulkNode(levels int, parts ...[]BulkItem) *node {
-	root := &node{
-		numEntries: len(parts),
-		parent:     nil,
-		isLeaf:     false,
-	}
+	root := nodePool.Get().(*node)
+	root.numEntries = len(parts)
 	for i, part := range parts {
 		child := bulkInsert(part, levels-1)
 		child.parent = root