--- a/rtree/insert.go
+++ b/rtree/insert.go
@@ -13,7 +13,8 @@ const (
 // Insert adds a new record to the RTree.
 func (t *RTree) Insert(box Box, recordID int) {
 	if t.root == nil {
-		t.root = &node{isLeaf: true}
+		t.root = nodePool.Get().(*node)
+		t.root.isLeaf = true
 	}
 
 	level := t.root.depth() - 1
@@ -49,7 +50,8 @@ func (t *RTree) adjustBoxesUpwards(node *node, box Box) {
 }
 
 func (t *RTree) joinRoots(r1, r2 *node) {
-	newRoot := &node{
+	newRoot := nodePool.Get().(*node)
+	*newRoot = node{
 		entries: [1 + maxChildren]entry{
 			entry{box: calculateBound(r1), child: r1},
 			entry{box: calculateBound(r2), child: r2},
@@ -140,7 +142,8 @@ func (t *RTree) splitNode(n *node) *node {
 
 	// Use the existing node for the 0 bits in the split, and a new node for
 	// the 1 bits in the split.
-	newNode := &node{isLeaf: n.isLeaf}
+	newNode := nodePool.Get().(*node)
+	*newNode = node{isLeaf: n.isLeaf}
 	totalEntries := n.numEntries
 	n.numEntries = 0
 	for i := 0; i < totalEntries; i++ {