--- a/rtree/perf_test.go
+++ b/rtree/perf_test.go
@@ -63,6 +63,7 @@ func BenchmarkInsert(b *testing.B) {
 				for j, b := range boxes {
 					tree.Insert(b, j)
 				}
+				tree.Recycle()
 			}
 		})
 	}