@peterstace 2020-06-18
I'm going to abandon this PR. While there are some decent performance gains reflected in the benchmark results, I have real concerns that due to the usage of `sync.Pool`, those same performance gains wouldn't be reflected in real life (`sync.Pool`'s performance characteristics are highly dependant of when GC runs, which isn't captured in a realistic way in the benchmarks).